Moved audience and grant type validation code to the request.go file
